requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
Django==4.1.7
django-jsonfield==1.4.1
django-model-utils==4.3.1
django-pwa==1.1.0
django-webpush==0.3.4
djangorestframework==3.14.0
http-ece==1.1.0
idna==3.4
jsonfield==3.1.0
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
py-vapid==1.9.0
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7.1
pywebpush==1.9.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tzdata==2022.7
urllib3==1.26.14
